rest tests: refactoring to have one central integration test. profiles added inorder for ContextLoadsTest to ignore configuration.
