+++ rest/src/test/java/tech/nermindedovic/rest/Topics.java
@@ -4,4 +4,5 @@ public class Topics {
     public static final String TRANSFER_OUTBOUND = "funds.transformer.request";
     public static final String BALANCE_OUTBOUND = "balance.transformer.request";
     public static final String BALANCE_INBOUND = "balance.transformer.response";
+    public static final String TRANSFER_ERROR = "funds.transfer.error";
 }
