+++ rest/src/test/java/tech/nermindedovic/rest/kafka/balance/BalanceTestConfig.java
@@ -2,10 +2,10 @@ package tech.nermindedovic.rest.kafka.balance;
 
 import io.confluent.kafka.schemaregistry.client.MockSchemaRegistryClient;
 import org.apache.kafka.clients.producer.ProducerConfig;
-import org.junit.jupiter.api.TestInstance;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Profile;
 import org.springframework.kafka.annotation.KafkaListener;
 import org.springframework.kafka.config.ConcurrentKafkaListenerContainerFactory;
 import org.springframework.kafka.core.DefaultKafkaConsumerFactory;
@@ -13,7 +13,6 @@ import org.springframework.kafka.core.DefaultKafkaProducerFactory;
 import org.springframework.kafka.core.KafkaTemplate;
 import org.springframework.kafka.support.serializer.JsonSerializer;
 import org.springframework.kafka.test.EmbeddedKafkaBroker;
-import org.springframework.kafka.test.context.EmbeddedKafka;
 import org.springframework.kafka.test.utils.KafkaTestUtils;
 import org.springframework.messaging.handler.annotation.SendTo;
 import tech.nermindedovic.AvroBalanceMessage;
@@ -26,8 +25,7 @@ import java.util.HashMap;
 import java.util.Map;
 
 @TestConfiguration
-@EmbeddedKafka(partitions = 1, topics = Topics.BALANCE_OUTBOUND, controlledShutdown = true)
-@TestInstance(TestInstance.Lifecycle.PER_CLASS)
+@Profile("integration")
 public class BalanceTestConfig {
 
     @Autowired
