+++ rest/src/test/java/tech/nermindedovic/rest/kafka/transfer/MockSerdeConfig.java
@@ -13,6 +13,7 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Primary;
+import org.springframework.context.annotation.Profile;
 import org.springframework.kafka.core.*;
 import org.springframework.kafka.listener.ContainerProperties;
 import org.springframework.kafka.listener.KafkaMessageListenerContainer;
@@ -28,6 +29,7 @@ import java.util.Map;
 import java.util.Properties;
 
 @TestConfiguration
+@Profile("integration")
 public class MockSerdeConfig {
